+++ examples/github_last.rb
@@ -19,9 +19,6 @@ if ENV['AUTH']
   headers['Authorization'] = "Basic #{auth}"
 end
 
-# Do not use deprecated indifferent access
-Restify::Processors::Json.indifferent_access = false
-
 gh    = Restify.new('https://api.github.com', headers:).get.value!
 user  = gh.rel(:user).get(user: 'jgraichen').value!
 repos = user.rel(:repos).get.value!