+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixes
 
 ### Breaks
+* Remove indifferent access methods (Hashie) from responses
 * Removed `em` and `em-pooled` adapters
 * Require Ruby 3.1+